1. Enable Network stack to pass SCT, currently MNP, ARP, IP4, TCP4 and DHCP4 have passed SCT. 2. Enable Network stack to pass ICC build. 3. Sync Network library instances' module names with their filenames.
git-svn-id: https://edk2.svn.sourceforge.net/svnroot/edk2/trunk@3692 6f19259b-4bc3-4df7-8a09-765794883524

+++ edk2/MdeModulePkg/Universal/Network/Ip4Dxe/Ip4Icmp.c
@@ -251,7 +251,7 @@ Ip4IcmpReplyEcho (
   Icmp                = (IP4_ICMP_QUERY_HEAD *) NetbufGetByte (Data, 0, NULL);
   Icmp->Head.Type     = ICMP_ECHO_REPLY;
   Icmp->Head.Checksum = 0;
-  Icmp->Head.Checksum = ~NetblockChecksum ((UINT8 *) Icmp, Data->TotalSize);
+  Icmp->Head.Checksum = (UINT16) (~NetblockChecksum ((UINT8 *) Icmp, Data->TotalSize));
 
   ReplyHead.Tos       = 0;
   ReplyHead.Fragment  = 0;
@@ -353,7 +353,7 @@ Ip4IcmpHandle (
     goto DROP;
   }
 
-  Checksum = ~NetbufChecksum (Packet);
+  Checksum = (UINT16) (~NetbufChecksum (Packet));
   if ((Icmp.Checksum != 0) && (Checksum != 0)) {
     goto DROP;
   }
